requirements.txt
customtkinter
pyserial
PyVirtualSerialPorts
